+++ setup/gcp/run_forseti_installer.py
@@ -18,22 +18,13 @@ This has been tested with python 2.7.
 """
 
 import argparse
-import getpass
-import pip
 
 from installer.forseti_server_installer import ForsetiServerInstaller
 from installer.forseti_client_installer import ForsetiClientInstaller
 
-INSTALLER_REQUIRED_PACKAGES = [
-    'ruamel.yaml'
-]
-
 
 def run():
     """Run the steps for the gcloud setup."""
-
-    # Install all the required packages
-    install_required_packages()
 
     parser = argparse.ArgumentParser()
     parser.add_argument('--no-cloudshell',
@@ -83,23 +74,6 @@ def run():
         forseti_setup = ForsetiClientInstaller(**args)
 
     forseti_setup.run_setup()
-
-
-def install(package_name):
-    """Install package.
-
-    Args:
-        package_name (str): Name of the package to install.
-    """
-    user = getpass.getuser()
-    pip.main(['install', package_name, '--user', user])
-
-
-def install_required_packages():
-    """Install required packages."""
-    for package in INSTALLER_REQUIRED_PACKAGES:
-        install(package)
-
 
 if __name__ == '__main__':
     run()
